Move avatar generation into main bot functions
Avatar can be loaded on connect, on a command, or on a random timer.
Only supports PNG/JPEG web APIs for now - SVG might be a good idea in future.
Avatar seed can be based on the bot ID, making each bot unique, or can be randomised each time it's fetched.